refactor: migrate formatters to TS
related to https://github.com/camunda/camunda/issues/22900
